Metadata for each field in the Colorado budget

Metadata for the many fields that exist in cobud_history or are calculated by cobud_forecast(). Manually assembled and maintained by the LCS economics team. See cobud - Colorado's Budget Forecast Tool and Budget Fields for more details.

Usage

cobud_field_metadata

Format

cobud_field_metadata

A data frame with 155 rows and 7 columns:

display_name

a nice readable name for the field

source

a hint at where this data can be located

category

a high level category: "revenue", "tabor", "expenditure", "context", or "other"

order

a sensible order to present the fields

download_group

a simple, user friendly grouping of the fields

needs_param

a logical indicating if it needs to be specified in the param when running cobud_forecast()

field_name

the column header in cobud_history

Source

LCS economics team

See also