Param templates

Each budget forecast, or call to cobud_forecast(), requires the user to specify the assumptions of the budget forecast run. These assumptions are specified in a nested list called a param. The param templates are examples of valid params.

See cobud - Colorado's Budget Forecast Tool and Budget Fields for more details.

Usage

detailed_param_template

simple_param_template

very_simple_param_template

Format

detailed_param_template

A nested list in the following format:

25-26

inflation_pct

population

26-27

inflation_pct

population

method

"Custom change"

value

1.05

adj

0.1

An object of class list of length 2.

An object of class list of length 2.

Source

https://content.leg.colorado.gov/EconomicForecasts

See also